В методе delete(Entity) добавлено existById(id)

+++ src/main/java/ru/zhadaev/schoolsecurity/service/GroupService.java
@@ -82,7 +82,11 @@ public class GroupService {
 
     @PreAuthorize("hasAuthority('GROUP_DELETE')")
     public void delete(Group group) {
-        groupRepository.delete(group);
+        if (existsById(group.getId())) {
+            groupRepository.delete(group);
+        } else {
+            throw new NotFoundException("Group delete error. Group not found");
+        }
     }
 
     @PreAuthorize("hasAuthority('GROUP_DELETE')")
